lesson 7 after review
