+++ src/by/teachmeskills/lesson7/task1Shapes/Square.java
@@ -1,0 +1,8 @@
+package by.teachmeskills.lesson7.task1Shapes;
+
+public class Square extends Rectangle {
+    public Square(double sideA) {
+        super(sideA, sideA);
+    }
+
+}
