+++ src/by/teachmeskills/lesson7/task1Shapes/Triangle.java
@@ -1,0 +1,22 @@
+package by.teachmeskills.lesson7.task1Shapes;
+
+public class Triangle extends Shape {
+    private final double sideA;
+    private final double sideB;
+    private final double sideC;
+
+    public Triangle(double sideA, double sideB, double sideC) {
+        this.sideA = sideA;
+        this.sideB = sideB;
+        this.sideC = sideC;
+    }
+
+    public double perimeter() {
+        return sideA + sideB + sideC;
+    }
+
+    public double area() {
+        double p = perimeter() / 2;
+        return Math.sqrt(p * (p - sideA) * (p - sideB) * (p - sideC));
+    }
+}
