+++ src/by/teachmeskills/lesson7/task2Jobs/Accountant.java
@@ -3,7 +3,7 @@ package by.teachmeskills.lesson7.task2Jobs;
 public class Accountant implements Print {
     private String activityName;
 
-    Accountant(String activityName) {
+    public Accountant(String activityName) {
         this.activityName = activityName;
     }
