+++ src/by/teachmeskills/lesson7/task2Jobs/Ceo.java
@@ -3,7 +3,7 @@ package by.teachmeskills.lesson7.task2Jobs;
 public class Ceo implements Print {
     private String ceoPosition;
 
-    Ceo(String ceoPosition) {
+    public Ceo(String ceoPosition) {
         this.ceoPosition = ceoPosition;
     }
