+++ src/by/teachmeskills/lesson7/task2Jobs/Worker.java
@@ -3,7 +3,7 @@ package by.teachmeskills.lesson7.task2Jobs;
 public class Worker implements Print {
     private String jobName;
 
-    Worker(String jobName) {
+    public Worker(String jobName) {
         this.jobName = jobName;
     }
